+++ CHANGELOG.md
@@ -61,7 +61,6 @@
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
-- Fix audio recording attachments not paused when the app goes to the background or the screen is covered with another one. [#5685](https://github.com/GetStream/stream-chat-android/pull/5685)
 
 ### ⬆️ Improved
 
@@ -75,6 +74,7 @@
 ### 🐞 Fixed
 - Fix audio recording attachments not paused when the app goes to the background or the screen is covered with another one. [#5685](https://github.com/GetStream/stream-chat-android/pull/5685)
 - Not show deleted poll messages. [#5689](https://github.com/GetStream/stream-chat-android/pull/5689)
+- Fix "Thread reply" item shown in the message options menu for messages in a Thread. [#5683](https://github.com/GetStream/stream-chat-android/pull/5683)
 
 ### ⬆️ Improved
 
@@ -84,6 +84,7 @@
 - ⚠️Add `Message` as an argument of the `textStyle` lambda parameter of the `MessageTextFormatter.defaultFormatter` method to allow customizing the text style based on the `Message`. [#5688](https://github.com/GetStream/stream-chat-android/pull/5688)
 
 ### ⚠️ Changed
+- `defaultMessageOptionsState()` now accepts an `isInThread` flag to show/hide the "Thread reply" option. [#5683](https://github.com/GetStream/stream-chat-android/pull/5683)
 
 ### ❌ Removed